# openapi.json
{
  "components": {
    "schemas": {
...
      "AppendCreate": {
        "properties": {
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "derived_from": {
            "description": "hex CIDs of cross-session inputs: another session's head node(s) or DUTs",
            "items": {
              "type": "string"
            },
            "title": "Derived From",
            "type": "array"
          },
          "dut": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "hex CID of the DUT this round produced",
            "title": "Dut"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "prev": {
            "description": "hex CIDs of this node's intra-session causal predecessors",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Prev",
            "type": "array"
          },
          "step_id": {
            "title": "Step Id",
            "type": "string"
          },
          "transformation": {
            "title": "Transformation",
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "step_id",
          "transformation",
          "prev"
        ],
        "title": "AppendCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LineageRead": {
        "properties": {
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "cid": {
            "title": "Cid",
            "type": "string"
          },
          "derived_from": {
            "items": {
              "type": "string"
            },
            "title": "Derived From",
            "type": "array"
          },
          "dut": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dut"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "prev": {
            "items": {
              "type": "string"
            },
            "title": "Prev",
            "type": "array"
          },
          "session_uuid": {
            "format": "uuid",
            "title": "Session Uuid",
            "type": "string"
          },
          "step_id": {
            "title": "Step Id",
            "type": "string"
          },
          "transformation": {
            "title": "Transformation",
            "type": "string"
          }
        },
        "required": [
          "cid",
          "session_uuid",
          "actor_id",
          "step_id",
          "transformation",
          "occurred_at",
          "prev",
          "derived_from",
          "dut"
        ],
        "title": "LineageRead",
        "type": "object"
      },
...
      "RecordCreate": {
        "properties": {
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "agent_output": {
            "title": "Agent Output",
            "type": "string"
          },
          "agent_uuid": {
            "format": "uuid",
            "title": "Agent Uuid",
            "type": "string"
          },
          "artifact_cids": {
            "items": {
              "type": "string"
            },
            "title": "Artifact Cids",
            "type": "array"
          },
          "business_object_keys": {
            "items": {
              "type": "string"
            },
            "title": "Business Object Keys",
            "type": "array"
          },
          "config_cid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "hex CID of the config used; defaults to the agent's active one",
            "title": "Config Cid"
          },
          "input_context": {
            "title": "Input Context",
            "type": "string"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "prev": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "hex CIDs of intra-session predecessors; defaults to the session frontier",
            "title": "Prev"
          },
          "sequence": {
            "default": 0,
            "title": "Sequence",
            "type": "integer"
          },
          "span_id": {
            "description": "runtime span handle, assigned at span start",
            "title": "Span Id",
            "type": "string"
          },
          "step_id": {
            "title": "Step Id",
            "type": "string"
          },
          "transformation": {
            "title": "Transformation",
            "type": "string"
          },
          "used_sessions": {
            "description": "sessions whose results this round consumed; each one's frontier is linked in derived_from (head-at-handoff)",
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Used Sessions",
            "type": "array"
          }
        },
        "required": [
          "agent_uuid",
          "span_id",
          "input_context",
          "agent_output",
          "transformation",
          "actor_id",
          "step_id"
        ],
        "title": "RecordCreate",
        "type": "object"
      },
      "RecordRead": {
        "properties": {
          "dut": {
            "$ref": "#/components/schemas/DUTRead"
          },
          "node": {
            "$ref": "#/components/schemas/LineageRead"
          }
        },
        "required": [
          "dut",
          "node"
        ],
        "title": "RecordRead",
        "type": "object"
      },
      "SessionCreate": {
        "properties": {
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "session_uuid": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "stable session key; defaults to a new uuid4",
            "title": "Session Uuid"
          },
          "step_id": {
            "title": "Step Id",
            "type": "string"
          },
          "transformation": {
            "default": "session-start",
            "title": "Transformation",
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "step_id"
        ],
        "title": "SessionCreate",
        "type": "object"
      },
      "SessionRead": {
        "properties": {
          "frontier": {
            "items": {
              "type": "string"
            },
            "title": "Frontier",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/LineageRead"
            },
            "title": "Nodes",
            "type": "array"
          },
          "session_uuid": {
            "format": "uuid",
            "title": "Session Uuid",
            "type": "string"
          }
        },
        "required": [
          "session_uuid",
          "nodes",
          "frontier"
        ],
        "title": "SessionRead",
        "type": "object"
      },
      "SessionSummary": {
        "properties": {
          "frontier": {
            "items": {
              "type": "string"
            },
            "title": "Frontier",
            "type": "array"
          },
          "last_occurred_at": {
            "format": "date-time",
            "title": "Last Occurred At",
            "type": "string"
          },
          "nodes": {
            "title": "Nodes",
            "type": "integer"
          },
          "session_uuid": {
            "format": "uuid",
            "title": "Session Uuid",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "session_uuid",
          "nodes",
          "started_at",
          "last_occurred_at",
          "frontier"
        ],
        "title": "SessionSummary",
        "type": "object"
      },
      "TracebackStep": {
        "properties": {
          "agent_cid": {
            "title": "Agent Cid",
            "type": "string"
          },
          "agent_output": {
            "title": "Agent Output",
            "type": "string"
          },
          "config_cid": {
            "title": "Config Cid",
            "type": "string"
          },
          "dut_cid": {
            "title": "Dut Cid",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          }
        },
        "required": [
          "occurred_at",
          "dut_cid",
          "agent_output",
          "agent_cid",
          "config_cid"
        ],
        "title": "TracebackStep",
        "type": "object"
      },
...
      "WatermarkRead": {
        "properties": {
          "cid": {
            "title": "Cid",
            "type": "string"
          },
          "watermark": {
            "title": "Watermark",
            "type": "string"
          }
        },
        "required": [
          "cid",
          "watermark"
        ],
        "title": "WatermarkRead",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "aits",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/lineage/": {
      "get": {
        "description": "Enumerate sessions, most recently active first. Includes sub-sessions\n(e.g. subagent work); their parents reach them via ``derived_from``.",
        "operationId": "lineage_list_sessions",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  },
                  "title": "Response Lineage List Sessions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "lineage"
        ]
      },
      "post": {
        "operationId": "lineage_start_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Session",
        "tags": [
          "lineage"
        ]
      }
    },
    "/lineage/nodes/{cid_hex}/": {
      "get": {
        "operationId": "lineage_get_node",
        "parameters": [
          {
            "in": "path",
            "name": "cid_hex",
            "required": true,
            "schema": {
              "title": "Cid Hex",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Node",
        "tags": [
          "lineage"
        ]
      }
    },
    "/lineage/nodes/{cid_hex}/traceback/": {
      "get": {
        "description": "Every contributing DUT in the node's ancestry, ordered by event time,\neach carrying its agent + config attribution.",
        "operationId": "lineage_trace_back",
        "parameters": [
          {
            "in": "path",
            "name": "cid_hex",
            "required": true,
            "schema": {
              "title": "Cid Hex",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TracebackStep"
                  },
                  "title": "Response Lineage Trace Back",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trace Back",
        "tags": [
          "lineage"
        ]
      }
    },
    "/lineage/nodes/{cid_hex}/watermark/": {
      "get": {
        "description": "Render the flat P1 watermark string straight off a node: its\n``derived_from`` parents (each head's produced DUT when it has one), the\ntransformation, and the asserted event time.",
        "operationId": "lineage_watermark",
        "parameters": [
          {
            "in": "path",
            "name": "cid_hex",
            "required": true,
            "schema": {
              "title": "Cid Hex",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatermarkRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Watermark",
        "tags": [
          "lineage"
        ]
      }
    },
    "/lineage/{session_uuid}/": {
      "get": {
        "operationId": "lineage_get_session_nodes",
        "parameters": [
          {
            "in": "path",
            "name": "session_uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session Nodes",
        "tags": [
          "lineage"
        ]
      }
    },
    "/lineage/{session_uuid}/append/": {
      "post": {
        "operationId": "lineage_append",
        "parameters": [
          {
            "in": "path",
            "name": "session_uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append",
        "tags": [
          "lineage"
        ]
      }
    },
    "/lineage/{session_uuid}/record/": {
      "post": {
        "description": "Record one round of work: the DUT span and its watermark node, chained\nonto the session, in a single call.",
        "operationId": "lineage_record",
        "parameters": [
          {
            "in": "path",
            "name": "session_uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record",
        "tags": [
          "lineage"
        ]
      }
    },
...
  }
}